requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
requests
openai
